ИСправлены замечания к Pull-request. Исправлены тесты к DraftProtocol и DraftApproach.

+++ backend/src/main/kotlin/com/jetbrains/life_science/protocol/service/DraftProtocolServiceImpl.kt
@@ -50,7 +50,9 @@ class DraftProtocolServiceImpl(
     }
 
     override fun delete(draftProtocolId: Long) {
-        get(draftProtocolId)
+        if (!repository.existsById(draftProtocolId)) {
+            throw DraftProtocolNotFoundException("Draft protocol with id $draftProtocolId is not found")
+        }
         repository.deleteById(draftProtocolId)
     }
